// x-pack/platform/plugins/shared/alerting/server/rules_client/types.d.ts
export interface RulesClientContext {
    readonly logger: Logger;
    readonly getUserName: () => Promise<string | null>;
    readonly spaceId: string;
    readonly namespace?: string;
    readonly taskManager: TaskManagerStartContract;
    readonly unsecuredSavedObjectsClient: SavedObjectsClientContract;
    readonly authorization: AlertingAuthorization;
    readonly ruleTypeRegistry: RuleTypeRegistry;
    readonly minimumScheduleInterval: AlertingRulesConfig['minimumScheduleInterval'];
    readonly maxScheduledPerMinute: AlertingRulesConfig['maxScheduledPerMinute'];
    readonly minimumScheduleIntervalInMs: number;
    readonly createAPIKey: (name: string) => Promise<CreateAPIKeyResult>;
    readonly getActionsClient: () => Promise<ActionsClient>;
    readonly actionsAuthorization: ActionsAuthorization;
    readonly getEventLogClient: () => Promise<IEventLogClient>;
    readonly internalSavedObjectsRepository: ISavedObjectsRepository;
    readonly encryptedSavedObjectsClient: EncryptedSavedObjectsClient;
    readonly kibanaVersion: PluginInitializerContext['env']['packageInfo']['version'];
    readonly auditLogger?: AuditLogger;
    readonly eventLogger?: IEventLogger;
    readonly changeTrackingService?: IScopedChangeTrackingService;
    readonly fieldsToExcludeFromPublicApi: Array<keyof SanitizedRule>;
    readonly isAuthenticationTypeAPIKey: () => boolean;
    readonly getAuthenticationAPIKey: (name: string) => CreateAPIKeyResult;
    readonly cloneAPIKey: (name: string) => Promise<CreateAPIKeyResult>;
    readonly cloneApiKeysOnCreate?: boolean;
    readonly connectorAdapterRegistry: ConnectorAdapterRegistry;
    readonly getAlertIndicesAlias: GetAlertIndicesAlias;
    readonly alertsService: AlertsService | null;
    readonly backfillClient: BackfillClient;
    readonly isSystemAction: (actionId: string) => boolean;
    readonly uiSettings: UiSettingsServiceStart;
    readonly shouldGrantUiam?: boolean;
    readonly isServerless: boolean;
    readonly featureFlags: FeatureFlagsStart;
}
// ...
export type CreateAPIKeyResult = {
    apiKeysEnabled: false;
} | {
    apiKeysEnabled: true;
    result?: SecurityPluginGrantAPIKeyResult;
    uiamResult?: SecurityPluginGrantAPIKeyResult;
};